Reuse code between process and cluster agents
