+++ pkg/util/kubernetes/kubelet/kubelet_orchestrator.go
@@ -38,11 +38,11 @@ type KubeUtilInterface interface {
 	ListContainers() ([]*containers.Container, error)
 	IsAgentHostNetwork() (bool, error)
 	UpdateContainerMetrics(ctrList []*containers.Container) error
-	GetRawLocalPodList() ([]v1.Pod, error)
+	GetRawLocalPodList() ([]*v1.Pod, error)
 }
 
 // GetRawLocalPodList returns the unfiltered pod list from the kubelet
-func (ku *KubeUtil) GetRawLocalPodList() ([]v1.Pod, error) {
+func (ku *KubeUtil) GetRawLocalPodList() ([]*v1.Pod, error) {
 	data, code, err := ku.QueryKubelet(kubeletPodPath)
 
 	if err != nil {
@@ -60,8 +60,13 @@ func (ku *KubeUtil) GetRawLocalPodList() ([]v1.Pod, error) {
 	if !ok {
 		return nil, fmt.Errorf("pod list type assertion failed on %v", podListData)
 	}
+	// transform []v1.Pod in []*v1.Pod
+	pods := make([]*v1.Pod, 0, len(podList.Items))
+	for _, p := range podList.Items {
+		pods = append(pods, &p)
+	}
 
-	return podList.Items, nil
+	return pods, nil
 }
 
 // ComputeStatus is mostly copied from kubernetes to match what users see in kubectl
