+++ tasks/cluster_agent.py
@@ -20,6 +20,7 @@ DEFAULT_BUILD_TAGS = [
     "kubeapiserver",
     "clusterchecks",
     "secrets",
+    "orchestrator",
 ]
